+++ src/errors/UnhealthyApplianceError.ts
@@ -1,0 +1,1 @@
+export class UnhealthyApplianceError extends Error {}