+++ src/errors/index.ts
@@ -1,3 +1,4 @@
 export * from './AbstractImplementationError'
 export * from './ProcessingError'
+export * from './UnhealthyApplianceError'
 export * from './ValidationError'